requirements.txt
requests~=2.31.0
Flask~=2.2.5
Flask-Bootstrap~=3.3.7.1
Flask-Login~=0.6.3
SQLAlchemy~=2.0.25
Werkzeug~=3.0.0
pycryptodome~=3.23.0
python-dotenv~=1.1.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
Flask-Mail
redis==6.2.0
celery==5.5.3
Flask-Migrate==4.1.0
psycopg2-binary==2.9.11
gevent==25.9.1
flask-limiter==4.1.1
pandas==3.0.1
